Changed all int lines in config to float - Config would reset if someone attempted to turn int lines into floats from the json so to negate confusion, changed all instances to float - Decimals are ignored as the (int) attribute is used to make the floats int again so no rounding is used Replaced changed Darkness Rise ability in NightProwlerGoal - Now summons Blackflame pillars after a delay around in an aoe and gives speed ang strength to itself Modularized armor & armor effects out of ChaosSet and WitheredArmor into separate classes Hallowheart gives total Wither immunity instead of just at 50% HP EnhancedArkenplates Mirror ability triggers at under 50% HP instead of 33% Arkenplate gives Magic Resistance too at under 50% HP

+++ src/main/java/net/soulsweaponry/client/registry/KeyBindRegistry.java
@@ -110,7 +110,7 @@ public class KeyBindRegistry {
                     boolean accept = false;
                     if (effect && client.player.hasStatusEffect(EffectRegistry.MOON_HERALD) && !client.player.getItemCooldownManager().isCoolingDown(ItemRegistry.MOONSTONE_RING)) {
                         accept = true;
-                        client.player.getItemCooldownManager().set(ItemRegistry.MOONSTONE_RING, ConfigConstructor.moonlight_ring_projectile_cooldown);
+                        client.player.getItemCooldownManager().set(ItemRegistry.MOONSTONE_RING, (int) ConfigConstructor.moonlight_ring_projectile_cooldown);
                     } else if (melee || controller) {
                         for (Hand hand : Hand.values()) {
                             ItemStack stack = client.player.getStackInHand(hand);
