+++ src/main/java/net/soulsweaponry/items/bow/KrakenSlayer.java
@@ -21,7 +21,7 @@ public class KrakenSlayer extends ModdedBow {
     public KrakenSlayer(Settings settings, Supplier<Ingredient> repairIngredientSupplier) {
         super(settings, repairIngredientSupplier);
         this.addTooltipAbility(TooltipAbilities.FAST_PULL, TooltipAbilities.THIRD_SHOT);
-        this.configure(new RangedConfig(ConfigConstructor.kraken_slayer_pull_time_ticks, ConfigConstructor.kraken_slayer_damage, ConfigConstructor.kraken_slayer_max_velocity));
+        this.configure(new RangedConfig((int) ConfigConstructor.kraken_slayer_pull_time_ticks, ConfigConstructor.kraken_slayer_damage, ConfigConstructor.kraken_slayer_max_velocity));
     }
 
     @Override
